Struct ethkey::ExtendedPublic

pub struct ExtendedPublic { /* fields omitted */ }

Extended public key, allows deterministic derivation of subsequent keys.

Methods

impl ExtendedPublic

pub fn new(public: Public, chain_code: H256) -> Self

New extended public key from known parent and chain code

pub fn from_secret(secret: &ExtendedSecret) -> Result<Self, DerivationError>

Create new extended public key from known secret

pub fn derive<T>(&self, index: Derivation<T>) -> Result<Self, DerivationError> where
T: Label,

Derive new public key Operation is defined only for index belongs [0..2^31)

pub fn public(&self) -> &Public