Struct ethkey::ExtendedSecret

pub struct ExtendedSecret { /* fields omitted */ }

Extended secret key, allows deterministic derivation of subsequent keys.

Methods

impl ExtendedSecret

pub fn with_code(secret: Secret, chain_code: H256) -> ExtendedSecret

New extended key from given secret and chain code.

pub fn new_random(secret: Secret) -> ExtendedSecret

New extended key from given secret with the random chain code.

pub fn new(secret: Secret) -> ExtendedSecret

New extended key from given secret. Chain code will be derived from the secret itself (in a deterministic way).

pub fn derive<T>(&self, index: Derivation<T>) -> ExtendedSecret where
T: Label,

Derive new private key

pub fn as_raw(&self) -> &Secret

Private key component of the extended key.